pyproject.toml
[project]
name = "torboxtg"
version = "0.1.0"
description = "Telegram bot to debrid Terabox links using Torbox API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.7",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio-throttle>=1.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
